新增新增账单查询接口，优化标准化返回

+++ database/user.go
@@ -5,7 +5,7 @@ import "ledger/conf"
 var UserTableName = "user"
 
 type User struct {
-	Id       int    `json:"id" xorm:"pk autoincr"` //主键+自增
+	Id       int64  `json:"id" xorm:"pk autoincr"` //主键+自增
 	Name     string `json:"name"`
 	Password string `json:"password"`
 	Email    string `json:"email"`
